package/libbtfmw_h418: new package
U-boot SPL requires to have a valid libbtfmw-h418.a file available.
Indeed the libbtfmw is in charge of the early platform preparation.
This is part of the targetpack-20160812-3 tarball which can be found within
the starkl delivery available at https://bitbucket.org/4kopen/starkl/get/v1.0.3.tar.bz2

Also generate a "u-boot.fragment" which is used during uboot compilation