package/waffle: fix pkg_config_path value
Default path goes for $(HOST)/lib/pkgconfig.  This does not work for
latest waffle version since prefix in the pkg-config file points to the
host directory.  This results in the concatenation of $(STAGING_DIR)/xxx
and $(HOST_DIR)/... which ultimately resolve in a file not found.

Instead, override the default path to $(STAGING_DIR)/usr/share.

Signed-off-by: Raphael Gallais-Pou <rgallaispou@gmail.com>